refactor: split handlers.ts into domain-specific modules
Split the monolithic handlers.ts (900+ lines) into focused domain modules:
- src/handlers/types.ts - shared types (ToolResult, HandlerContext)
- src/handlers/memory.ts - store, get, list, update, delete, bulk ops
- src/handlers/recall.ts - recall, search, context, suggested
- src/handlers/relations.ts - relations CRUD, graph traversal
- src/handlers/admin.ts - status, init, ingest, export, migrate, etc.
- src/handlers/index.ts - router that delegates to domain handlers

Original handlers.ts re-exports for backward compatibility.

Closes #60
