+++ src/handlers/memory.ts
@@ -1,0 +1,366 @@
+import { formatMemory, withConcurrency, validateContentLength, validateImportance, UPDATE_FIELDS } from '../format.js';
+import type { HandlerContext, ToolResult } from './types.js';
+
+export async function handleMemory(ctx: HandlerContext, name: string, args: any): Promise<ToolResult | null> {
+  const { makeRequest } = ctx;
+
+  switch (name) {
+    case 'memoclaw_store': {
+      const { content, importance, tags, namespace, memory_type, session_id, agent_id, expires_at, pinned, immutable } = args;
+      if (!content || (typeof content === 'string' && content.trim() === '')) {
+        throw new Error('content is required and cannot be empty');
+      }
+      validateContentLength(content);
+      validateImportance(importance);
+      const body: any = { content };
+      if (importance !== undefined) body.importance = importance;
+      if (tags) body.tags = tags;
+      if (namespace) body.namespace = namespace;
+      if (memory_type) body.memory_type = memory_type;
+      if (session_id) body.session_id = session_id;
+      if (agent_id) body.agent_id = agent_id;
+      if (expires_at) body.expires_at = expires_at;
+      if (pinned !== undefined) body.pinned = pinned;
+      if (immutable !== undefined) body.immutable = immutable;
+      const result = await makeRequest('POST', '/v1/store', body);
+      return { content: [{ type: 'text', text: `✅ Memory stored\n${formatMemory(result.memory || result)}\n\n${JSON.stringify(result, null, 2)}` }] };
+    }
+
+    case 'memoclaw_get': {
+      const { id } = args;
+      if (!id) throw new Error('id is required');
+      const result = await makeRequest('GET', `/v1/memories/${id}`);
+      return { content: [{ type: 'text', text: `${formatMemory(result.memory || result)}\n\n${JSON.stringify(result, null, 2)}` }] };
+    }
+
+    case 'memoclaw_list': {
+      const { limit, offset, tags, namespace, memory_type, session_id, agent_id, after } = args;
+      const params = new URLSearchParams();
+      if (limit !== undefined) params.set('limit', String(limit));
+      if (offset !== undefined) params.set('offset', String(offset));
+      if (namespace) params.set('namespace', namespace);
+      if (memory_type) params.set('memory_type', memory_type);
+      if (tags && Array.isArray(tags) && tags.length > 0) params.set('tags', tags.join(','));
+      if (session_id) params.set('session_id', session_id);
+      if (agent_id) params.set('agent_id', agent_id);
+      if (after) params.set('after', after);
+      const result = await makeRequest('GET', `/v1/memories?${params}`);
+      const memories = result.memories || result.data || [];
+      const total = result.total ?? memories.length;
+      const formatted = memories.length > 0
+        ? '\n\n' + memories.map((m: any) => formatMemory(m)).join('\n\n')
+        : '';
+      return { content: [{ type: 'text', text: `Showing ${memories.length} of ${total} memories${formatted}\n\n---\n${JSON.stringify(result, null, 2)}` }] };
+    }
+
+    case 'memoclaw_update': {
+      const { id, ...allFields } = args;
+      if (!id) throw new Error('id is required');
+      const updateFields: Record<string, any> = {};
+      for (const [key, value] of Object.entries(allFields)) {
+        if (UPDATE_FIELDS.has(key) && value !== undefined) updateFields[key] = value;
+      }
+      if (Object.keys(updateFields).length === 0) {
+        throw new Error('No valid update fields provided. Allowed: ' + [...UPDATE_FIELDS].join(', '));
+      }
+      if (typeof updateFields.content === 'string') validateContentLength(updateFields.content);
+      validateImportance(updateFields.importance);
+      const result = await makeRequest('PATCH', `/v1/memories/${id}`, updateFields);
+      return { content: [{ type: 'text', text: `✅ Memory ${id} updated\n${formatMemory(result.memory || result)}\n\n${JSON.stringify(result, null, 2)}` }] };
+    }
+
+    case 'memoclaw_delete': {
+      const { id } = args;
+      if (!id) throw new Error('id is required');
+      const result = await makeRequest('DELETE', `/v1/memories/${id}`);
+      return { content: [{ type: 'text', text: `🗑️ Memory ${id} deleted\n\n${JSON.stringify(result, null, 2)}` }] };
+    }
+
+    case 'memoclaw_bulk_delete': {
+      const { ids } = args;
+      if (!ids || !Array.isArray(ids) || ids.length === 0) {
+        throw new Error('ids is required and must be a non-empty array');
+      }
+      if (ids.length > 100) throw new Error('Maximum 100 IDs per bulk delete call');
+
+      // Try dedicated bulk-delete endpoint first, fall back to one-by-one
+      let succeeded = 0;
+      let failed = 0;
+      let errors: string[] = [];
+      try {
+        const result = await makeRequest('POST', '/v1/memories/bulk-delete', { ids });
+        succeeded = result.deleted ?? ids.length;
+        failed = (result.failed && result.failed.length) || 0;
+        if (result.failed && result.failed.length > 0) {
+          errors = result.failed.map((f: any) => `${f.id}: ${f.error || 'unknown error'}`);
+        }
+      } catch {
+        // Fallback: delete one-by-one
+        const results = await withConcurrency(
+          ids.map((id: string) => () => makeRequest('DELETE', `/v1/memories/${id}`)),
+          10
+        );
+        succeeded = results.filter(r => r.status === 'fulfilled').length;
+        failed = results.filter(r => r.status === 'rejected').length;
+        errors = results
+          .map((r, i) => r.status === 'rejected' ? `${ids[i]}: ${(r as PromiseRejectedResult).reason?.message || 'unknown error'}` : null)
+          .filter(Boolean) as string[];
+      }
+
+      let text = `🗑️ Bulk delete: ${succeeded} succeeded, ${failed} failed`;
+      if (errors.length > 0) text += `\n\nErrors:\n${errors.join('\n')}`;
+      return { content: [{ type: 'text', text }] };
+    }
+
+    case 'memoclaw_bulk_store': {
+      const { memories, session_id, agent_id } = args;
+      if (!memories || !Array.isArray(memories) || memories.length === 0) {
+        throw new Error('memories is required and must be a non-empty array');
+      }
+      if (memories.length > 100) throw new Error('Maximum 100 memories per bulk store call');
+      for (const [i, m] of memories.entries()) {
+        if (!m.content || (typeof m.content === 'string' && m.content.trim() === '')) {
+          throw new Error(`Memory at index ${i} has empty content`);
+        }
+        validateContentLength(m.content, `Memory at index ${i}`);
+        validateImportance(m.importance, `Memory at index ${i} importance`);
+      }
+      const STORE_FIELDS = ['content', 'importance', 'tags', 'namespace', 'memory_type', 'pinned', 'expires_at', 'immutable'];
+
+      // Try batch API endpoint first (single request), fall back to one-by-one
+      try {
+        const batchBody: any = {
+          memories: memories.map((m: any) => {
+            const item: any = {};
+            for (const key of STORE_FIELDS) {
+              if (m[key] !== undefined) item[key] = m[key];
+            }
+            if (session_id) item.session_id = session_id;
+            if (agent_id) item.agent_id = agent_id;
+            return item;
+          }),
+        };
+        const result = await makeRequest('POST', '/v1/store/batch', batchBody);
+        const stored = result.memories || result.data || [];
+        const failedItems = result.failed || [];
+        let text = `✅ Bulk store: ${stored.length} stored, ${failedItems.length} failed`;
+        if (stored.length > 0) text += `\n\n${stored.map((m: any) => formatMemory(m)).join('\n\n')}`;
+        if (failedItems.length > 0) {
+          const errors = failedItems.map((f: any) => `index ${f.index ?? '?'}: ${f.error || 'unknown error'}`);
+          text += `\n\nErrors:\n${errors.join('\n')}`;
+        }
+        return { content: [{ type: 'text', text }] };
+      } catch (batchErr: any) {
+        // Fall back to one-by-one if batch endpoint is unavailable (404)
+        if (!batchErr.message?.includes('404') && !batchErr.message?.includes('Not Found')) {
+          throw batchErr;
+        }
+      }
+
+      // Fallback: store one-by-one with concurrency
+      const results = await withConcurrency(
+        memories.map((m: any) => () => {
+          const body: any = {};
+          for (const key of STORE_FIELDS) {
+            if (m[key] !== undefined) body[key] = m[key];
+          }
+          if (session_id) body.session_id = session_id;
+          if (agent_id) body.agent_id = agent_id;
+          return makeRequest('POST', '/v1/store', body);
+        }),
+        10
+      );
+      const succeeded = results.filter(r => r.status === 'fulfilled');
+      const failed = results.filter(r => r.status === 'rejected');
+      const stored = succeeded.map(r => (r as PromiseFulfilledResult<any>).value?.memory || (r as PromiseFulfilledResult<any>).value);
+      const errors = failed.map((r) => {
+        const idx = results.indexOf(r);
+        return `index ${idx}: ${(r as PromiseRejectedResult).reason?.message || 'unknown error'}`;
+      });
+      let text = `✅ Bulk store: ${succeeded.length} stored, ${failed.length} failed`;
+      if (stored.length > 0) text += `\n\n${stored.map((m: any) => formatMemory(m)).join('\n\n')}`;
+      if (errors.length > 0) text += `\n\nErrors:\n${errors.join('\n')}`;
+      return { content: [{ type: 'text', text }] };
+    }
+
+    case 'memoclaw_import': {
+      const { memories, session_id, agent_id } = args;
+      if (!memories || !Array.isArray(memories) || memories.length === 0) {
+        throw new Error('memories is required and must be a non-empty array');
+      }
+      if (memories.length > 100) throw new Error('Maximum 100 memories per import call');
+      for (const [i, m] of memories.entries()) {
+        if (!m.content || (typeof m.content === 'string' && m.content.trim() === '')) {
+          throw new Error(`Memory at index ${i} has empty content`);
+        }
+        validateContentLength(m.content, `Memory at index ${i}`);
+        validateImportance(m.importance, `Memory at index ${i} importance`);
+      }
+
+      const IMPORT_FIELDS = ['content', 'importance', 'tags', 'namespace', 'memory_type', 'pinned', 'immutable'];
+
+      // Try batch API endpoint first
+      try {
+        const batchBody: any = {
+          memories: memories.map((m: any) => {
+            const item: any = {};
+            for (const key of IMPORT_FIELDS) {
+              if (m[key] !== undefined) item[key] = m[key];
+            }
+            if (session_id) item.session_id = session_id;
+            if (agent_id) item.agent_id = agent_id;
+            return item;
+          }),
+        };
+        const result = await makeRequest('POST', '/v1/store/batch', batchBody);
+        const stored = result.memories || result.data || [];
+        const failedItems = result.failed || [];
+        let text = `📥 Import: ${stored.length} stored, ${failedItems.length} failed`;
+        if (failedItems.length > 0) {
+          const errors = failedItems.map((f: any) => `index ${f.index ?? '?'}: ${f.error || 'unknown error'}`);
+          text += `\n\nErrors:\n${errors.join('\n')}`;
+        }
+        return { content: [{ type: 'text', text }] };
+      } catch (batchErr: any) {
+        if (!batchErr.message?.includes('404') && !batchErr.message?.includes('Not Found')) {
+          throw batchErr;
+        }
+      }
+
+      // Fallback: store one-by-one with concurrency
+      const results = await withConcurrency(
+        memories.map((m: any) => () => {
+          const body: any = { content: m.content };
+          if (m.importance !== undefined) body.importance = m.importance;
+          if (m.tags) body.tags = m.tags;
+          if (m.namespace) body.namespace = m.namespace;
+          if (m.memory_type) body.memory_type = m.memory_type;
+          if (m.pinned !== undefined) body.pinned = m.pinned;
+          if (m.immutable !== undefined) body.immutable = m.immutable;
+          if (session_id) body.session_id = session_id;
+          if (agent_id) body.agent_id = agent_id;
+          return makeRequest('POST', '/v1/store', body);
+        }),
+        10
+      );
+      const succeeded = results.filter(r => r.status === 'fulfilled').length;
+      const failed = results.filter(r => r.status === 'rejected').length;
+      const errors = results
+        .map((r, i) => r.status === 'rejected' ? `index ${i}: ${(r as PromiseRejectedResult).reason?.message || 'unknown error'}` : null)
+        .filter(Boolean);
+      let text = `📥 Import: ${succeeded} stored, ${failed} failed`;
+      if (errors.length > 0) text += `\n\nErrors:\n${errors.join('\n')}`;
+      return { content: [{ type: 'text', text }] };
+    }
+
+    case 'memoclaw_pin': {
+      const { id } = args;
+      if (!id) throw new Error('id is required');
+      const result = await makeRequest('PATCH', `/v1/memories/${id}`, { pinned: true });
+      return { content: [{ type: 'text', text: `📌 Memory ${id} pinned\n${formatMemory(result.memory || result)}` }] };
+    }
+
+    case 'memoclaw_unpin': {
+      const { id } = args;
+      if (!id) throw new Error('id is required');
+      const result = await makeRequest('PATCH', `/v1/memories/${id}`, { pinned: false });
+      return { content: [{ type: 'text', text: `📌 Memory ${id} unpinned\n${formatMemory(result.memory || result)}` }] };
+    }
+
+    case 'memoclaw_batch_update': {
+      const { updates } = args;
+      if (!updates || !Array.isArray(updates) || updates.length === 0) {
+        throw new Error('updates is required and must be a non-empty array');
+      }
+      if (updates.length > 50) throw new Error('Maximum 50 updates per batch update call');
+      for (const [i, u] of updates.entries()) {
+        if (!u.id) throw new Error(`Update at index ${i} is missing "id"`);
+        validateImportance(u.importance, `Update at index ${i} importance`);
+      }
+      try {
+        const result = await makeRequest('POST', '/v1/memories/batch-update', { updates });
+        const updated = result.updated ?? result.memories?.length ?? '?';
+        const memories = result.memories || [];
+        let text = `✅ Batch update: ${updated} memories updated`;
+        if (memories.length > 0) text += `\n\n${memories.map((m: any) => formatMemory(m)).join('\n\n')}`;
+        return { content: [{ type: 'text', text: `${text}\n\n${JSON.stringify(result, null, 2)}` }] };
+      } catch (err: any) {
+        if (err.message?.includes('404') || err.message?.includes('Not Found')) {
+          const results = await withConcurrency(
+            updates.map((u: any) => () => {
+              const { id, ...fields } = u;
+              const updateFields: Record<string, any> = {};
+              for (const [key, value] of Object.entries(fields)) {
+                if (UPDATE_FIELDS.has(key) && value !== undefined) updateFields[key] = value;
+              }
+              return makeRequest('PATCH', `/v1/memories/${id}`, updateFields);
+            }),
+            10
+          );
+          const succeeded = results.filter(r => r.status === 'fulfilled');
+          const failed = results.filter(r => r.status === 'rejected');
+          const memories = succeeded.map(r => (r as PromiseFulfilledResult<any>).value?.memory || (r as PromiseFulfilledResult<any>).value);
+          const errors = failed.map((r) => {
+            const idx = results.indexOf(r);
+            return `${updates[idx]?.id}: ${(r as PromiseRejectedResult).reason?.message || 'unknown error'}`;
+          });
+          let text = `✅ Batch update: ${succeeded.length} updated, ${failed.length} failed`;
+          if (memories.length > 0) text += `\n\n${memories.map((m: any) => formatMemory(m)).join('\n\n')}`;
+          if (errors.length > 0) text += `\n\nErrors:\n${errors.join('\n')}`;
+          return { content: [{ type: 'text', text }] };
+        }
+        throw err;
+      }
+    }
+
+    case 'memoclaw_count': {
+      const { namespace, tags, agent_id, memory_type } = args;
+      const params = new URLSearchParams();
+      if (namespace) params.set('namespace', namespace);
+      if (tags && Array.isArray(tags) && tags.length > 0) params.set('tags', tags.join(','));
+      if (agent_id) params.set('agent_id', agent_id);
+      if (memory_type) params.set('memory_type', memory_type);
+
+      let total: number | string = 'unknown';
+      try {
+        const countResult = await makeRequest('GET', `/v1/memories/count?${params}`);
+        total = countResult.count ?? countResult.total ?? 'unknown';
+      } catch {
+        params.set('limit', '1');
+        params.set('offset', '0');
+        const result = await makeRequest('GET', `/v1/memories?${params}`);
+        if (typeof result.total === 'number') {
+          total = result.total;
+        } else {
+          const memories = result.memories || result.data || [];
+          if (memories.length === 0) {
+            total = 0;
+          } else {
+            let counted = 0;
+            let offset = 0;
+            const pageSize = 100;
+            while (offset < 100000) {
+              const pageParams = new URLSearchParams(params);
+              pageParams.set('limit', String(pageSize));
+              pageParams.set('offset', String(offset));
+              const page = await makeRequest('GET', `/v1/memories?${pageParams}`);
+              const items = page.memories || page.data || [];
+              counted += items.length;
+              if (typeof page.total === 'number') { total = page.total; break; }
+              if (items.length < pageSize) { total = counted; break; }
+              offset += pageSize;
+            }
+            if (total === 'unknown') total = `${counted}+`;
+          }
+        }
+      }
+
+      const filters = [namespace && `namespace=${namespace}`, memory_type && `type=${memory_type}`, agent_id && `agent=${agent_id}`, tags?.length && `tags=${tags.join(',')}`].filter(Boolean);
+      const filterStr = filters.length > 0 ? ` (${filters.join(', ')})` : '';
+      return { content: [{ type: 'text', text: `📊 Total memories${filterStr}: ${total}` }] };
+    }
+
+    default:
+      return null;
+  }
+}
